+++ test/functional/helpers/consoleLogger/index.js
