+++ test/functional/specs/desktop/C2585.js
@@ -1,6 +1,6 @@
 import { Selector } from "testcafe";
-import testServerUrl from "../../src/constants/testServerUrl";
-import fixtureFactory from "../../src/fixtureFactory";
+import testServerUrl from "../../helpers/constants/testServerUrl";
+import fixtureFactory from "../../helpers/fixtureFactory";
 
 const urlCollector = `${testServerUrl}/test/functional/sandbox/html/alloySdk.html`;
