+++ test/functional/specs/desktop/C2586.js
@@ -1,5 +1,5 @@
 import { Selector } from "testcafe";
-import fixtureFactory from "../../src/fixtureFactory";
+import fixtureFactory from "../../helpers/fixtureFactory";
 
 const urlCollector = `http://127.0.0.1:8080/test/functional/sandbox/html/alloySdk.html?alloy_debug=true`;
