+++ test/functional/specs/desktop/C2588.js
@@ -1,4 +1,4 @@
-import fixtureFactory from "../../src/fixtureFactory";
+import fixtureFactory from "../../helpers/fixtureFactory";
 
 const urlCollector = `http://127.0.0.1:8080/test/functional/sandbox/html/multiConfig.html`;
